third_party/blink/web_tests/external/wpt/editing/other/typing-around-link-element-after-joining-paragraphs_action=Backspace-expected.txt

This is a testharness.js-based test.
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a><br></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a><br></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)">foo</a><br><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span></p>" but got "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span></p>" but got "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span></p>"
[PASS] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
  assert_equals: expected "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span></p>" but got "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span></p>"
[FAIL] Backspace in "<p><a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
  assert_equals: expected "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span></p>" but got "<p><a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span></p>"
[FAIL] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b></a><br><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span>" but got "<a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span>"
[FAIL] Backspace in "<a href="about:blank" style="color:rgb(0, 0, 255)"><b>foo</b><br></a><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
  assert_equals: expected "<a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a><b>XY</b><span style=\"color:rgb(255, 0, 0)\">bar</span>" but got "<a href=\"about:blank\" style=\"color:rgb(0, 0, 255)\"><b>foo</b></a>XY<span style=\"color:rgb(255, 0, 0)\">bar</span>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></b></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></b><br></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a></b></p><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></b></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></b><br></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<p><b><a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a></b></p><span style="color:rgb(255, 0, 0)">[]bar</span>"
[PASS] Backspace in "<b><a href="about:blank" style="color:rgb(0, 0, 255)">foo</a></b><br><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
[PASS] Backspace in "<b><a href="about:blank" style="color:rgb(0, 0, 255)">foo<br></a></b><p><span style="color:rgb(255, 0, 0)">[]bar</span></p>"
Harness: the test ran to completion.
